aider: Add alert when "Open Chest" button is clicked, handle loot action by removing item from room and changing button color to 'none'.
# Aider chat conversation:

USER: Let's make it so when i click on open chest, I get an alert with the cotents of the chest, with the options - Loot or Leave. also, if i take the loot, we need to remove the item from the room as well, and perhaps change the button from warning to none.

ASSISTANT: To implement this feature, we need to modify the following files:

1. `src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.ts`
2. `src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.vue`

Here are the changes we need to make:

1. In `TempleGrounds.ts`, we need to modify the `roomActions` computed property to add an alert when the "Open Chest" button is clicked. We also need to add a method to handle the loot action, which will remove the item from the room and change the button color.

2. In `TempleGrounds.vue`, we need to bind the button color to a dynamic property so it can change based on the room state.

Here are the detailed changes:

```typescript
src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.ts
<<<<<<< HEAD
          actions.buttons = [{
            text: "Open Chest",
            handler: () => {
              // handle loot
            }
          }]
=======
          actions.buttons = [{
            text: "Open Chest",
            handler: () => {
              alert(`You found a ${this.currentRoom.content.item}!`);
              const action = confirm('Do you want to take the loot?');
              if (action) {
                this.handleLoot();
              }
            }
          }]
>>>>>>> updated
```

```typescript
src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.ts
<<<<<<< HEAD
  methods: {
    clickFight() {
      alert()
    },
=======
  methods: {
    handleLoot() {
      // Remove the item from the room
      delete this.currentRoom.content;
      // Change the button color
      this.currentRoom.type = 'empty';
    },
    clickFight() {
      alert()
    },
>>>>>>> updated
```

```vue
src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.vue
<<<<<<< HEAD
      <ion-fab-button :color="actionColor">
=======
      <ion-fab-button :color="currentRoom.content ? actionColor : 'none'">
>>>>>>> updated
        <i class="fad fa-2x" :class="ROOM_ICONS[currentRoom.type]" />
      </ion-fab-button>
```

These changes will add an alert when the "Open Chest" button is clicked, showing the contents of the chest. If the user chooses to take the loot, the item will be removed from the room and the button color will change to 'none'.