aider: Added computed property `validRoomCoords` that returns an array of valid room coordinates.
# Aider chat conversation:

ASSISTANT: To create a computed property that returns an array of all valid room coordinates that aren't walls, we need to modify the file `src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.ts`.

We will add a new computed property `validRoomCoords` that will iterate over the `maze` array and check the `type` of each room in the `rooms` object. If the `type` is not equal to `"wall"`, it will add the coordinates to the array in the format `"X,Y"`.

Here is the *edit block* for the change:

```typescript
src/views/MyPortal/HomeTown/TempleGrounds/TempleGrounds.ts
<<<<<<< HEAD
    canMoveRight() {
      const [row, col] = this.currentPosition;
      if (col >= this.maze[0].length - 1) return false;
      return this.rooms[this.maze[row][col + 1]].type !== "wall";
    },
=======
    canMoveRight() {
      const [row, col] = this.currentPosition;
      if (col >= this.maze[0].length - 1) return false;
      return this.rooms[this.maze[row][col + 1]].type !== "wall";
    },
    validRoomCoords() {
      let validCoords = [];
      for (let row = 0; row < this.maze.length; row++) {
        for (let col = 0; col < this.maze[row].length; col++) {
          if (this.rooms[this.maze[row][col]].type !== "wall") {
            validCoords.push(`[${row},${col}]`);
          }
        }
      }
      return validCoords;
    },
>>>>>>> updated
```

This new computed property `validRoomCoords` will return an array of strings, each string representing the coordinates of a valid room in the format `"X,Y"`. You can use this array in your template to list the `backgroundImage` for each valid room.